Hoan thien command 1

+++ source/count_comparison.cpp
@@ -2,9 +2,9 @@
 
 using namespace std;
 
-int selectionSort_count_cmp(int arr[], int n)
+long long selectionSort_count_cmp(int arr[], int n)
 {
-    int count_comparison = 0;
+    long long count_comparison = 0;
     for (int i = 0; i < n - 1; i++)
     {
         count_comparison++;
@@ -20,9 +20,9 @@ int selectionSort_count_cmp(int arr[], int n)
     return count_comparison;
 }
 
-int shakerSort_count_cmp(int arr[], int n)
+long long shakerSort_count_cmp(int arr[], int n)
 {
-    int count_comparison = 0;
+    long long count_comparison = 0;
     int left = 0;
     int right = n - 1;
     int k = 0;
@@ -52,7 +52,7 @@ int shakerSort_count_cmp(int arr[], int n)
     return count_comparison;
 }
 
-int sortFirstMiddleLast_count_cmp(int arr[], int first, int last, int &count_comparison)
+int sortFirstMiddleLast_count_cmp(int arr[], int first, int last, long long &count_comparison)
 {
     int mid = (first + last) / 2;
     if (++count_comparison && arr[first] > arr[mid])
@@ -64,7 +64,7 @@ int sortFirstMiddleLast_count_cmp(int arr[], int first, int last, int &count_com
     return mid;
 }
 
-int partition_count_cmp(int arr[], int first, int last, int &count_comparison)
+int partition_count_cmp(int arr[], int first, int last, long long &count_comparison)
 {
     int pivotIndex = sortFirstMiddleLast_count_cmp(arr, first, last, count_comparison);
     swap(arr[pivotIndex], arr[last - 1]);
@@ -94,7 +94,7 @@ int partition_count_cmp(int arr[], int first, int last, int &count_comparison)
     return indexFromLeft;
 }
 
-void quickSort_count_cmp(int arr[], int first, int last, int &count_comparison)
+void quickSort_count_cmp(int arr[], int first, int last, long long &count_comparison)
 {
     if (last - first == 1)
     {
@@ -114,40 +114,51 @@ void quickSort_count_cmp(int arr[], int first, int last, int &count_comparison)
 
 // ------------------------
 
-void insertionSort_count_cmp(int arr[], int n, long& count_comparison) {
-    for (int i = 1;i < n;i++, ++count_comparison) {
+void insertionSort_count_cmp(int arr[], int n, long long &count_comparison)
+{
+    for (int i = 1; i < n; i++, ++count_comparison)
+    {
         int numKeep = arr[i];
-        for (int j = i;j >= 1;j--, ++count_comparison) {
-            if (arr[j - 1] > numKeep) {
+        for (int j = i; j >= 1; j--, ++count_comparison)
+        {
+            if (arr[j - 1] > numKeep)
+            {
                 ++count_comparison;
                 arr[j] = arr[j - 1];
                 arr[j - 1] = numKeep;
             }
-            else break;
+            else
+                break;
         }
     }
 }
 
-void heapify_count_cmp(int arr[], int end, int curNode, long& count_comparison) {
+void heapify_count_cmp(int arr[], int end, int curNode, long long &count_comparison)
+{
     int leftNode = curNode * 2 + 1, rightNode = curNode * 2 + 2;
     int bigNode = curNode;
 
-    if (arr[leftNode] > arr[bigNode]) {
+    if (arr[leftNode] > arr[bigNode])
+    {
         count_comparison++;
-        if (leftNode < end) {
+        if (leftNode < end)
+        {
             count_comparison++;
             bigNode = leftNode;
         }
     }
-    if (arr[rightNode] > arr[bigNode]) {
+    if (arr[rightNode] > arr[bigNode])
+    {
         count_comparison++;
-        if (rightNode < end) {
+        if (rightNode < end)
+        {
             count_comparison++;
             bigNode = rightNode;
         }
     }
 
-    if (bigNode != curNode) {
+    if (bigNode != curNode)
+    {
         swap(arr[curNode], arr[bigNode]);
         heapify_count_cmp(arr, end, bigNode, count_comparison);
     }
@@ -155,38 +166,48 @@ void heapify_count_cmp(int arr[], int end, int curNode, long& count_comparison) 
     count_comparison++;
 }
 
-void heapSort_count_cmp(int arr[], int n, long& count_comparison) {
-    for (int i = (n - 1) / 2;i >= 0;i--, ++count_comparison) heapify_count_cmp(arr, n, i, count_comparison);
-    for (int i = n - 1;i > 0;i--, ++count_comparison) {
+void heapSort_count_cmp(int arr[], int n, long long &count_comparison)
+{
+    for (int i = (n - 1) / 2; i >= 0; i--, ++count_comparison)
+        heapify_count_cmp(arr, n, i, count_comparison);
+    for (int i = n - 1; i > 0; i--, ++count_comparison)
+    {
         swap(arr[0], arr[i]);
         heapify_count_cmp(arr, i, 0, count_comparison);
     }
 }
 
-void radixSort_count_cmp(int arr[], int n, long& count_comparison) {
+void radixSort_count_cmp(int arr[], int n, long long &count_comparison)
+{
     int maxNum = arr[0];
-    for (int i = 1;i < n;i++, ++count_comparison) {
+    for (int i = 1; i < n; i++, ++count_comparison)
+    {
         maxNum = max(maxNum, arr[i]);
         ++count_comparison;
     }
-    for (int exp = 1;maxNum / exp > 0;exp *= 10, ++count_comparison) {
-        int* storage[10]; int cnt[10];
-        for (int i = 0;i < 10;i++, ++count_comparison) {
+    for (int exp = 1; maxNum / exp > 0; exp *= 10, ++count_comparison)
+    {
+        int *storage[10];
+        int cnt[10];
+        for (int i = 0; i < 10; i++, ++count_comparison)
+        {
             storage[i] = new int[n];
             cnt[i] = 0;
         }
 
-        for (int i = 0;i < n;i++, ++count_comparison) {
+        for (int i = 0; i < n; i++, ++count_comparison)
+        {
             int idx = (arr[i] / exp) % 10;
             storage[idx][cnt[idx]++] = arr[i];
         }
 
         int idx = 0;
-        for (int i = 0;i < 10;i++, ++count_comparison) {
-            for (int j = 0;j < cnt[i];j++, ++count_comparison) {
+        for (int i = 0; i < 10; i++, ++count_comparison)
+        {
+            for (int j = 0; j < cnt[i]; j++, ++count_comparison)
+            {
                 arr[idx++] = storage[i][j];
             }
         }
     }
 }
-
