+++ source/count_comparison.h
@@ -1,17 +1,17 @@
 #include <iostream>
 #include <algorithm>
 
-int selectionSort_count_cmp(int arr[], int n);
+long long selectionSort_count_cmp(int arr[], int n);
 
-int shakerSort_count_cmp(int arr[], int n);
+long long shakerSort_count_cmp(int arr[], int n);
 
-int sortFirstMiddleLast_count_cmp(int arr[], int first, int last, int &count_comparison);
-int partition_count_cmp(int arr[], int first, int last, int &count_comparison);
-void quickSort_count_cmp(int arr[], int first, int last, int &count_comparison);
+int sortFirstMiddleLast_count_cmp(int arr[], int first, int last, long long &count_comparison);
+int partition_count_cmp(int arr[], int first, int last, long long &count_comparison);
+void quickSort_count_cmp(int arr[], int first, int last, long long &count_comparison);
 
-void insertionSort_count_cmp(int arr[], int n, long& count_comparison);
+void insertionSort_count_cmp(int arr[], int n, long long &count_comparison);
 
-void heapify_count_cmp(int arr[], int end, int curNode, long& count_comparison);
-void heapSort_count_cmp(int arr[], int n, long& count_comparison);
+void heapify_count_cmp(int arr[], int end, int curNode, long long &count_comparison);
+void heapSort_count_cmp(int arr[], int n, long long &count_comparison);
 
-void radixSort_count_cmp(int arr[], int n, long& count_comparison);
+void radixSort_count_cmp(int arr[], int n, long long &count_comparison);
